requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
faiss-cpu==1.8.0
sentence-transformers==3.0.1
streamlit==1.39.0
python-dotenv==1.0.1
pypdf2==3.0.1
pymupdf==1.24.5
python-multipart==0.0.9
requests==2.32.4